+++ include/treaty.h
@@ -40,7 +40,7 @@
 struct trtstr {
     /* initial part must match struct empobj */
     short ef_type;
-    short trt_uid;
+    int trt_uid;
     unsigned trt_seqno;
     time_t trt_timestamp;
     /* end of part matching struct empobj */
